Fix TypeError Carbon::addDays recibe string en vez de int
Config::value() retorna string desde BD. Carbon en producción requiere int|float.
Castear a (int) en las 3 ocurrencias de addDays(Config::value('dias_vigencia_token')).

+++ app/Http/Controllers/Admin/OrdenesController.php
@@ -681,7 +681,7 @@ class OrdenesController extends Controller
                 'direccion' => $evaluadoData['direccion'] ?? null,
                 'orden_id' => $orden->id,
                 'token_unico' => Str::random(32),
-                'token_expira_at' => now()->addDays(\App\Models\Config::value('dias_vigencia_token') ?? 30),
+                'token_expira_at' => now()->addDays((int) (\App\Models\Config::value('dias_vigencia_token') ?? 30)),
                 // Nuevos campos granulares
                 'tipo_servicio' => $evaluadoData['tipo_servicio'],
                 // Regla de negocio: socioeconómico siempre usa formulario preempleo
@@ -889,7 +889,7 @@ class OrdenesController extends Controller
             // Regenerar token si expiró
             $evaluado->update([
                 'token_unico' => EvaluadoOrden::generarToken(),
-                'token_expira_at' => now()->addDays(\App\Models\Config::value('dias_vigencia_token') ?? 30),
+                'token_expira_at' => now()->addDays((int) (\App\Models\Config::value('dias_vigencia_token') ?? 30)),
             ]);
             $evaluado->refresh();
         }
@@ -1108,7 +1108,7 @@ class OrdenesController extends Controller
                 'completado_at'              => null,
                 'estado_evaluacion'          => 'pendiente',
                 'token_unico'                => EvaluadoOrden::generarToken(),
-                'token_expira_at'            => now()->addDays(\App\Models\Config::value('dias_vigencia_token') ?? 30),
+                'token_expira_at'            => now()->addDays((int) (\App\Models\Config::value('dias_vigencia_token') ?? 30)),
             ]);
 
             DB::commit();
